New - regenerate Koji repos for external repo changes
The Kojira service only appears to concern itself with issuing
a regen-repo command whenever builds within Koji indicate that it's
internal repository meta-data needs to be refreshed.  This revision
establishes a cron job that monitors the external repositories for
changes and issues a regen-repo command for the build-tag when needed.