New - mash support in Module[koji]
This extends the Koji module with support for mash, including a script
to mash a package repository for each of many Koji tags.  These are then
composited together into a cohesive whole and then synchronized with
a published set, ready for consumption by tools such as yum and dnf.
Additionally, a repoview page is generated for each.