New - deployment of Sigul auto_signer
Finally this has come together for doubledog and all atop CentOS 7!
Most everything here is pretty straightforward, but one item especially
deserves attention.  The Koji Hub gains another Common Name for which
its allowed to proxy.  In this case the Sigul Bridge has been added.
This was necessary because the Sigul Bridge was logging an error message
from Koji stating:

    AuthError: CN=arche-f23.doubledog.org,OU=Sigul
    Bridge,O=doubledog.org,L=Mason,ST=Michigan,C=US is not authorized to
    login other users