Change - koji::regen_repos now uses koji-helpers package
Class[koji::regen_repos] has been heavily revised here to eliminate the
regen-repos script, the template from which it originated, the cron job
to execute it, etc.  Instead, the class now relies on the newly
developed koji-helpers package which provides a rewritten regen-repos
tool that is now available as a system service.