Change - koji::mash now uses koji-helpers package
Class[koji::mash] has been heavily revised here to eliminate the
mash-everything script, the template from which it originated, the cron
job to execute it, etc.  Instead, the class now relies on the newly
developed koji-helpers package which provides a rewritten
mash-everything tool that is now available as a system service.

This also revises parts of commit cf8d91d1 (for doing the same with
regen-repos) as certain parts conflicted with the above and to improve
symmetry along with minor corrections.