New - koji::mash::repo::debug_info_path parameter
In playing with the mash tool and non-default paths, especially those
that had some depth, it was observed the debug/ directory was still
placed at the top.  A bit of digging in the code revealed the
"debuginfo_path" setting even though none of the shipped example
configuration files employ it.  Testing verified that it indeed does as
expected so it has been added with full support here.