Bug - koji-gc: Values instance has no attribute 'krb_rdns'
koji-utils-1.10.1-10.el7.noarch has introduced a bug in that there's
apparently no internal default for a new setting named 'krb_rdns'.
Details of this can be found at:
https://fedorahosted.org/koji/ticket/328

At this time there doesn't appear to be any official bug report beyond
comment #20 in the above.  Nevertheless, this revision provides a value
for the setting so that the GC can continue working as before.