New - Koji CLI profile support
Class[koji::cli] has been refactored to extract a new
Define[koji::cli::profile].  Parameters to the latter are effectively
handle the same as with the former, however there is now of course
a requirement for namevar so that each profile can be uniquely
identified.  Meanwhile, Class[koji::cli] replaces all those parameters
with a single one named "profiles" which expects a hash suitable for
create_resources thereby allowing the entire set of parameters for all
profiles to neatly arrive via Hiera, if so desired.