Bug - Koji clients using wrong CA for the Hub
The configuration file for the Koji GC merely specifies that this should
be the "certificate of the CA that issued the HTTP server certificate",
but upon further reflection I'm almost 100% certain that this must be
the CA of the Hub not the Web.  The Hub does serve XML-RPC over HTTP so
I think this just a poorly written comment.  All other Koji clients (of
which the GC is just another) specify the CA of the Hub so it only
stands to reason this case would be the same.

This also corrects a subtle error with the GC and some (though not all)
other clients in that were using the Hub's client CA certificate rather
than the Hub's CA certificate.  In our deployment, they're the same but
if they weren't, I'm fairly certain this would be a bug.