New - koji::hub::web_url parameter
This new parameter, which defaults to "http://FQDN/koji" - the value
used in prior versions, allows you to override the KojiWebURL setting in
the /etc/koji/hub.conf file.  While the default may suffice for many
cases, there may be situations where the Hub's canonical FQDN is not
ideal such as when a better alias might be required for access through
a reverse proxy allowing Internet access.  This also allows you to
specify HTTPS instead of the default HTTP should that be desired.

Signed-off-by: John Florian <jflorian@doubledog.org>