Bug - kojid fails rpm builds on Fedora 30
Fedora 30 still has kojid running under Python 2, but python2-librepo is
no longer available.  There is a python3-librepo, but for whatever
reason the Fedora Project is not yet leveraging the Python 3 support for
the Builder.  Use of yum appears to be the least favored option, but
I see no alternative at this point.

Signed-off-by: John Florian <jflorian@doubledog.org>